+++ database/migrations/2026_04_12_000003_create_contract_installments_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->unsignedBigInteger('status_label_id');
             $table->string('ticket_reference', 100)->nullable();
             $table->text('notes')->nullable();
-            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
             $table->timestamps();
             $table->softDeletes();
