+++ database/migrations/2026_04_12_000004_create_contract_amendments_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->date('effective_date');
             $table->string('ticket_reference', 100)->nullable();
             $table->text('notes')->nullable();
-            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
             $table->timestamps();
             $table->softDeletes();
